+++ NEWS.md
@@ -2,6 +2,7 @@
 # ==============================
 * Improved documentation of datasets included in the package
 * Added smarter warning message for tryCatch
+* Tweak code_male handling
 
 # ggpedigree 1.1.0.0
 * Saved raw excel data as csv to have a non-proprietary format for data storage